Netflix catalogue in C#
This version is incomplete, it imports and exports data, it has the main menu, and it lists all titles from data file.

+++ NetflixCatalogueSystem/Episode.cs
@@ -11,6 +11,6 @@ namespace NetflixCatalogueSystem
         public int number;
         public string name;
         public string duration;
-        public decimal rating;
+        public double rating;
     }
 }
